refactor: change implementations/models -> implementations/segmentators
