--- auto_ml/implementations/models/__init__.py
@@ -1,13 +1,0 @@
-"""Model implementations subpackage."""
-
-from auto_ml.implementations.models.base import InMemoryPyTorchDataset
-from auto_ml.implementations.models.quadtree_model import QuadtreeSegmentationModel
-from auto_ml.implementations.models.swin_model import SwinModel
-from auto_ml.implementations.models.vit_model import ViTModel
-
-__all__ = [
-    "InMemoryPyTorchDataset",
-    "ViTModel",
-    "SwinModel",
-    "QuadtreeSegmentationModel",
-]
