+++ auto_ml/implementations/segmentators/base.py
