+++ auto_ml/implementations/segmentators/quadtree_model.py
