+++ auto_ml/implementations/segmentators/vit_model.py
@@ -8,7 +8,7 @@ import torch.nn as nn
 import torch.optim as optim
 from torch.utils.data import DataLoader
 
-from auto_ml.implementations.models.base import InMemoryPyTorchDataset
+from auto_ml.implementations.segmentators.base import InMemoryPyTorchDataset
 from auto_ml.interfaces import (
     DatasetInterface,
     MaskPair,
